Rearranged directory to have a folder containing all B models
