--- B4/README.md
@@ -1,1 +1,0 @@
-This folder contains the relevant documents and code relating to Subgroup B question 4, Cost-Effectiveness of Campaigns
