+++ GrpB_models/B4/README.md
@@ -1,0 +1,1 @@
+This folder contains the relevant documents and code relating to Subgroup B question 4, Cost-Effectiveness of Campaigns
